These are the new features added to the program:
Display of index numbers in the ListView for easy reference of text entries.

Drag and drop file feature for quick loading of "global-metadata.dat" file.

Convenient export and import of text feature, which allows users to modify exported text content using a text editor and view the changes in the ListView after importing.